Add data directory in Dockerfile and remove directory creation from wechat.go
In this commit, a new directory named 'data' is created in the Dockerfile. This
change simplifies the setup process by ensuring the 'data' directory is present
in the Docker container. Additionally, the code for creating the 'data'
directory in 'bootstrap/wechat.go' has been removed as it is no longer
necessary.